+++ frontend/dashboard/app/llms-full.txt/route.ts
@@ -1,4 +1,4 @@
-import { generateLlmsFullTxt } from "@/app/utils/llms/markdown_generator";
+import { generateLlmsFullTxt } from "@/app/utils/llms/llms_full_txt";
 
 // Prerendered at build time: the docs tree, the blog posts and the
 // marketing page.md twins are all fixed at build.
